# change way the workers close out