--- a/worker.go
+++ b/worker.go
@@ -1,5 +1,9 @@
 package bigopool
 
+import (
+	"sync"
+)
+
 type (
 	Worker struct {
 		// A channel for receiving a job that was dispatched
@@ -23,8 +27,10 @@ func NewWorker(jobCh chan Job, errCh chan error, resultCh chan Result) Worker {
 
 // Start method starts the run loop for the worker, listening for a quit channel in
 // case we need to stop it
-func (w Worker) Start() {
+func (w Worker) Start(wg *sync.WaitGroup) {
+	wg.Add(1)
 	go func() {
+		defer wg.Done()
 		for {
 			select {
 			case job, more := <-w.jobCh: