We were seeing a very odd bug where the WaitGroup in the Dispatcher would occasionally panic going negative. I still don't know how it's possible but it was happening 4-5 times a day and killing our service.

```
panic: sync: negative WaitGroup counter    
goroutine 7064575 [running]:  
sync.(*WaitGroup).Add(0xc00013e540, 0xffffffffffffffff)       
/.gimme/versions/go1.16.13.linux.amd64/src/sync/waitgroup.go:74 +0x147       
sync.(*WaitGroup).Done(...)       
/.gimme/versions/go1.16.13.linux.amd64/src/sync/waitgroup.go:99   
github.com/bigodines/bigopool.(*Dispatcher).Run.func1(0xc000404070)     
/gopath/pkg/mod/github.com/bigodines/bigopool@v3.2.0+incompatible/pool.go:105 +0x14d   
created by github.com/bigodines/bigopool.(*Dispatcher).Run      
/gopath/pkg/mod/github.com/bigodines/bigopool@v3.2.0+incompatible/pool.go:97 +0x75 
```

To avoid this I changed up the way that the pool finishes up the job by starting the process in the Dispatcher.Wait() to close the worker's job channel so they start closing out. Once they close out then we wait for the Dispatcher.Run() to close.

Surprising to me the benchmark seems to be on par with the original method of waiting and letting the worker go routines close out after everything.

I ran each benchmark three times with original vs this branch and took the best result from each. Maybe I got lucky with virus scan not running for the feature branch.

original benchmark
```
go test -bench=.  -benchmem=true -cpu=1,2,4,8 -run Bench
goos: darwin
goarch: amd64
pkg: github.com/bigodines/bigopool
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
Benchmark1Workers1Queue           	 2179256	       539.2 ns/op	      99 B/op	       0 allocs/op
Benchmark1Workers1Queue-2         	 1915218	       616.7 ns/op	      90 B/op	       0 allocs/op
Benchmark1Workers1Queue-4         	 1817332	       631.5 ns/op	      95 B/op	       0 allocs/op
Benchmark1Workers1Queue-8         	 1904367	       653.2 ns/op	      90 B/op	       0 allocs/op
Benchmark5Workers1000Queue        	 2671977	       411.0 ns/op	      80 B/op	       0 allocs/op
Benchmark5Workers1000Queue-2      	 2392610	       495.8 ns/op	      90 B/op	       0 allocs/op
Benchmark5Workers1000Queue-4      	 2194755	       539.6 ns/op	      98 B/op	       0 allocs/op
Benchmark5Workers1000Queue-8      	 2174530	       534.5 ns/op	      99 B/op	       0 allocs/op
Benchmark10Workers100Queue        	 2668249	       420.7 ns/op	      80 B/op	       0 allocs/op
Benchmark10Workers100Queue-2      	 2464814	       481.3 ns/op	      87 B/op	       0 allocs/op
Benchmark10Workers100Queue-4      	 2313861	       507.6 ns/op	      93 B/op	       0 allocs/op
Benchmark10Workers100Queue-8      	 2094547	       546.4 ns/op	      82 B/op	       0 allocs/op
Benchmark20Workers200Queue        	 2671730	       422.2 ns/op	      80 B/op	       0 allocs/op
Benchmark20Workers200Queue-2      	 2502262	       482.3 ns/op	      86 B/op	       0 allocs/op
Benchmark20Workers200Queue-4      	 2408499	       499.3 ns/op	      89 B/op	       0 allocs/op
Benchmark20Workers200Queue-8      	 2206192	       551.4 ns/op	      97 B/op	       0 allocs/op
Benchmark20Workers10000Queue      	 2662543	       424.9 ns/op	      81 B/op	       0 allocs/op
Benchmark20Workers10000Queue-2    	 2439786	       466.5 ns/op	      88 B/op	       0 allocs/op
Benchmark20Workers10000Queue-4    	 2424310	       493.6 ns/op	      89 B/op	       0 allocs/op
Benchmark20Workers10000Queue-8    	 2169270	       532.9 ns/op	      99 B/op	       0 allocs/op
Benchmark100Workers10000Queue     	 2661888	       421.9 ns/op	      81 B/op	       0 allocs/op
Benchmark100Workers10000Queue-2   	 2726989	       470.5 ns/op	      99 B/op	       0 allocs/op
Benchmark100Workers10000Queue-4   	 2419137	       497.4 ns/op	      89 B/op	       0 allocs/op
Benchmark100Workers10000Queue-8   	 2212276	       535.2 ns/op	      97 B/op	       0 allocs/op
PASS
ok  	github.com/bigodines/bigopool	42.053s
```

this branch
```
go test -bench=.  -benchmem=true -cpu=1,2,4,8 -run Bench
goos: darwin
goarch: amd64
pkg: github.com/bigodines/bigopool
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
Benchmark1Workers1Queue           	 2239099	       561.8 ns/op	      96 B/op	       0 allocs/op
Benchmark1Workers1Queue-2         	 2068951	       577.8 ns/op	      83 B/op	       0 allocs/op
Benchmark1Workers1Queue-4         	 1944374	       605.0 ns/op	      88 B/op	       0 allocs/op
Benchmark1Workers1Queue-8         	 1889241	       649.5 ns/op	      91 B/op	       0 allocs/op
Benchmark5Workers1000Queue        	 2766888	       441.2 ns/op	      97 B/op	       0 allocs/op
Benchmark5Workers1000Queue-2      	 2524478	       466.6 ns/op	      85 B/op	       0 allocs/op
Benchmark5Workers1000Queue-4      	 2200882	       518.8 ns/op	      98 B/op	       0 allocs/op
Benchmark5Workers1000Queue-8      	 2274174	       539.7 ns/op	      95 B/op	       0 allocs/op
Benchmark10Workers100Queue        	 2702775	       409.5 ns/op	      79 B/op	       0 allocs/op
Benchmark10Workers100Queue-2      	 2482053	       479.2 ns/op	      87 B/op	       0 allocs/op
Benchmark10Workers100Queue-4      	 2305369	       521.5 ns/op	      93 B/op	       0 allocs/op
Benchmark10Workers100Queue-8      	 2195844	       544.8 ns/op	      98 B/op	       0 allocs/op
Benchmark20Workers200Queue        	 2627191	       440.5 ns/op	      82 B/op	       0 allocs/op
Benchmark20Workers200Queue-2      	 2520817	       469.1 ns/op	      85 B/op	       0 allocs/op
Benchmark20Workers200Queue-4      	 2337250	       504.8 ns/op	      92 B/op	       0 allocs/op
Benchmark20Workers200Queue-8      	 2177000	       557.4 ns/op	      99 B/op	       0 allocs/op
Benchmark20Workers10000Queue      	 2708474	       402.5 ns/op	      79 B/op	       0 allocs/op
Benchmark20Workers10000Queue-2    	 2493572	       465.3 ns/op	      86 B/op	       0 allocs/op
Benchmark20Workers10000Queue-4    	 2345312	       542.8 ns/op	      92 B/op	       0 allocs/op
Benchmark20Workers10000Queue-8    	 2188532	       555.0 ns/op	      98 B/op	       0 allocs/op
Benchmark100Workers10000Queue     	 2526268	       408.4 ns/op	      85 B/op	       0 allocs/op
Benchmark100Workers10000Queue-2   	 2632839	       452.7 ns/op	      82 B/op	       0 allocs/op
Benchmark100Workers10000Queue-4   	 2398586	       491.9 ns/op	      90 B/op	       0 allocs/op
Benchmark100Workers10000Queue-8   	 2225059	       545.6 ns/op	      97 B/op	       0 allocs/op
PASS
ok  	github.com/bigodines/bigopool	41.130s
```